Rename TypicalPersons to TypicalClients
