+++ src/test/java/seedu/address/ui/ClientListPanelTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTimeoutPreemptively;
 import static seedu.address.testutil.EventsUtil.postNow;
 import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
-import static seedu.address.testutil.TypicalPersons.getTypicalPersons;
+import static seedu.address.testutil.TypicalClients.getTypicalClients;
 import static seedu.address.ui.testutil.GuiTestAssert.assertCardDisplaysPerson;
 import static seedu.address.ui.testutil.GuiTestAssert.assertCardEquals;
 
@@ -26,7 +26,7 @@ import seedu.address.storage.XmlSerializableAddressBook;
 
 public class ClientListPanelTest extends GuiUnitTest {
     private static final ObservableList<Contact> TYPICAL_CONTACTS =
-            FXCollections.observableList(getTypicalPersons());
+            FXCollections.observableList(getTypicalClients());
 
     private static final JumpToListRequestEvent JUMP_TO_SECOND_EVENT = new JumpToListRequestEvent(INDEX_SECOND_PERSON);
